Adds regtest client to run integration test in nakamotoJs
