+++ packages/nakamotojs/src/index.js
@@ -2,19 +2,10 @@ import * as address from './address.js';
 import * as crypto from './crypto.js';
 import * as networks from './networks.js';
 import * as payments from './payments/index.js';
-import * as RegtestClient from './regtest_client.js';
 import * as script from './script.js';
 import * as bip371 from './psbt/bip371.js';
 import * as bufferUtils from './bufferutils.js';
-export {
-  address,
-  crypto,
-  networks,
-  payments,
-  RegtestClient,
-  script,
-  bufferUtils,
-};
+export { address, crypto, networks, payments, script, bufferUtils };
 export { Block } from './block.js';
 export { Psbt } from './psbt.js';
 export { OPS as opcodes } from './ops.js';
