+++ packages/nakamotojs/test/integration/csv.spec.ts
@@ -15,10 +15,10 @@ import {
 // @ts-ignore
 import * as bip68 from 'bip68';
 import * as varuint from 'varuint-bitcoin';
-import { RegtestClient } from '../../src/regtest_client.js';
-import { getRandomAddress } from '../../src/utils.js';
+import { getRandomAddress } from '../test_utils.js';
+import { RegtestClient } from './regtest_client.js';
 import { Transaction } from '../../src/index.js';
-import { CHAIN, NETWORK } from '../../src/config/index.js';
+import { CHAIN, NETWORK } from './config/index.js';
 import { getNetwork } from '../../src/networks.js';
 
 const ECPair = ECPairFactory(ecc);
