+++ packages/nakamotojs/test/integration/regtest_client.ts
@@ -1,7 +1,9 @@
+/// <reference path="../../bitcoind-rpc.d.ts" />
+
 import RpcClient from 'bitcoind-rpc';
 import * as util from 'util';
 import { ECPairFactory } from 'ecpair';
-import { rpcJSON2CB } from './utils.js';
+import { rpcJSON2CB, _Transaction } from '../../src/utils.js';
 import * as ecc from '@bitcoin-computer/secp256k1';
 import {
   RPC_HOST,
@@ -12,36 +14,41 @@ import {
   CHAIN,
   NETWORK,
 } from './config/index.js';
-import { payments, Psbt, networks } from './index.js';
+
+import { payments, Psbt, networks } from '../../src/index.js';
+
 const network = networks.getNetwork(CHAIN, NETWORK);
 const ECPair = ECPairFactory(ecc);
+
 const rpcConfig = {
   protocol: RPC_PROTOCOL,
   user: RPC_USER,
   pass: RPC_PASSWORD,
   host: RPC_HOST,
-  port: parseInt(RPC_PORT, 10),
+  port: parseInt(RPC_PORT!, 10),
 };
+
 const rpcClientObj = new RpcClient(rpcConfig);
-const nativeRpcClient = {};
+const nativeRpcClient: any = {};
 const callSpec = JSON.parse(JSON.stringify(RpcClient.callspec));
 Object.keys(callSpec).forEach(spec => {
   callSpec[spec.toLowerCase()] = callSpec[spec];
 });
+
 const types = {
-  str(arg) {
+  str(arg: any) {
     return arg.toString();
   },
-  string(arg) {
+  string(arg: any) {
     return arg.toString();
   },
-  int(arg) {
+  int(arg: any) {
     return parseFloat(arg);
   },
-  float(arg) {
+  float(arg: any) {
     return parseFloat(arg);
   },
-  bool(arg) {
+  bool(arg: any) {
     return (
       arg === true ||
       arg === '1' ||
@@ -50,14 +57,15 @@ const types = {
       arg.toString().toLowerCase() === 'true'
     );
   },
-  obj(arg) {
+  obj(arg: any) {
     if (typeof arg === 'string') {
       return JSON.parse(arg);
     }
     return arg;
   },
 };
-export function parseParams(method, params) {
+
+export function parseParams(method: string, params: string): any[] {
   const callSpecMethod = callSpec[method];
   if (!callSpecMethod) {
     throw new Error('This RPC method does not exist or is not supported');
@@ -75,6 +83,7 @@ export function parseParams(method, params) {
     return types[type](param);
   });
 }
+
 try {
   // methods like add node
   Object.keys(RpcClient.prototype).forEach(method => {
@@ -88,49 +97,74 @@ try {
       );
     }
   });
-} catch (error) {
+} catch (error: any) {
   throw new Error(`Error occurred while binding RPC methods: ${error.message}`);
 }
+
 export class RegtestClient {
+  nativeRpcClient: any;
+
   constructor() {
     this.nativeRpcClient = nativeRpcClient;
   }
-  async faucet(address, value) {
+
+  async faucet(
+    address: string,
+    value: number,
+  ): Promise<{
+    txId: string;
+    vout: number;
+    height: number;
+    satoshis: number;
+  }> {
     // Fund address
     const valueBtc = value / 1e8;
     const { result: txId } = await nativeRpcClient.sendtoaddress(
       address,
       valueBtc,
     );
+
+    const randomAddress = await nativeRpcClient.getnewaddress();
+
     // Mine block on top
-    await nativeRpcClient.generateToAddress(
-      1,
-      'mvFeNF9DAR7WMuCpBPbKuTtheihLyxzj8i',
-    );
+    await nativeRpcClient.generateToAddress(1, randomAddress.result);
     const { result: fetchedTx } = await nativeRpcClient.getrawtransaction(
       txId,
       1,
     );
+
     // Prepare return value
-    const vout = fetchedTx.vout.findIndex(x => x.value * 1e8 === value);
+    const vout = fetchedTx.vout.findIndex(
+      (x: { value: number }) => x.value * 1e8 === value,
+    );
+
     return { txId, vout, height: -1, satoshis: value };
   }
-  async mine(count) {
+
+  async mine(count: number): Promise<any> {
     const { result: address } = await nativeRpcClient.getnewaddress();
     return await nativeRpcClient.generatetoaddress(count, address);
   }
-  async broadcast(txHex) {
+
+  async broadcast(txHex: string): Promise<string> {
     const { result, error } = await nativeRpcClient.sendRawTransaction(txHex);
     if (error) {
       throw new Error('Error sending transaction');
     }
     return result;
   }
-  async getTx(txId) {
+
+  async getTx(txId: string): Promise<_Transaction> {
     const res = await nativeRpcClient.getRawTransaction(txId, 1);
     return rpcJSON2CB(res.result);
   }
-  async verify(txo) {
+
+  async verify(txo: {
+    txId: string;
+    vout: number;
+    address?: string;
+    satoshis?: number;
+  }): Promise<void> {
     const tx = await this.getTx(txo.txId);
     const txoActual = tx.outs[txo.vout];
     if (txo.address && txoActual.address !== txo.address)
@@ -138,22 +172,37 @@ export class RegtestClient {
     if (txo.satoshis && txoActual.value !== txo.satoshis)
       throw new Error('Value Mismatch');
   }
-  async height() {
+
+  async height(): Promise<number> {
     const { result: topBlockHash } = await nativeRpcClient.getbestblockhash();
+
     const { result } = await nativeRpcClient.getblockheader(topBlockHash, true);
     return result.height;
   }
-  async faucetScript(script, value) {
+
+  async faucetScript(
+    script: Buffer,
+    value: number,
+  ): Promise<{
+    txId: string;
+    vout: number;
+    height: number;
+    satoshis: number;
+  }> {
     const key = ECPair.makeRandom({ network });
     const payment = payments.p2pkh({
       pubkey: key.publicKey,
       network,
     });
     const { address } = payment;
+
     const txId = (
       await nativeRpcClient.sendtoaddress(address, (value * 2) / 1e8, '', '')
     ).result;
-    const fetchedTx = (await nativeRpcClient.getrawtransaction(txId, 1)).result;
+
+    const fetchedTx: any = (await nativeRpcClient.getrawtransaction(txId, 1))
+      .result;
+
     // Find the output index (vout) corresponding to the address
     let voutIndex = -1;
     for (let i = 0; i < fetchedTx.vout.length; i++) {
@@ -167,18 +216,22 @@ export class RegtestClient {
         break;
       }
     }
+
     if (voutIndex === -1) {
       throw new Error('Could not find the output index for the given address');
     }
+
     // Verify the UTXO exists using getTxOut
     let counter = 10;
     let unspent;
     do {
       unspent = await nativeRpcClient.getTxOut(txId, voutIndex, true);
     } while (unspent.error && counter--);
+
     if (unspent.error) {
       throw new Error('Could not find the faucet transaction');
     }
+
     const txvb = new Psbt({ network });
     txvb.addInput({
       hash: txId,
@@ -191,10 +244,12 @@ export class RegtestClient {
     });
     txvb.signInput(0, key);
     txvb.finalizeAllInputs();
+
     const txv = txvb.extractTransaction(true);
     const { result: txIdScript } = await nativeRpcClient.sendrawtransaction(
       txv.toHex(),
     );
+
     // Check the vout
     let voutPsbt = 0;
     let foundPsbt = await nativeRpcClient.getTxOut(txIdScript, voutPsbt, true);
@@ -203,6 +258,7 @@ export class RegtestClient {
     if (foundPsbt.error)
       foundPsbt = await nativeRpcClient.getTxOut(txIdScript, ++voutPsbt, true);
     if (foundPsbt.error) throw new Error('No UTXO found');
+
     return {
       txId: txv.getId(),
       vout: voutPsbt,
