+++ packages/nakamotojs/test/integration/transactions.spec.ts
@@ -11,10 +11,10 @@ import {
   opcodes,
 } from '@bitcoin-computer/nakamotojs';
 
-import { RegtestClient } from '../../src/regtest_client.js';
-import { getRandomAddress } from '../../src/utils.js';
+import { getRandomAddress } from '../test_utils.js';
+import { RegtestClient } from './regtest_client.js';
 import { Transaction } from '../../src/index.js';
-import { CHAIN, NETWORK } from '../../src/config/index.js';
+import { CHAIN, NETWORK } from './config/index.js';
 import { getNetwork } from '../../src/networks.js';
 
 export const fail = () => {
