+++ packages/nakamotojs/ts_src/index.ts
@@ -2,20 +2,11 @@ import * as address from './address.js';
 import * as crypto from './crypto.js';
 import * as networks from './networks.js';
 import * as payments from './payments/index.js';
-import * as RegtestClient from './regtest_client.js';
 import * as script from './script.js';
 import * as bip371 from './psbt/bip371.js';
 import * as bufferUtils from './bufferutils.js';
 
-export {
-  address,
-  crypto,
-  networks,
-  payments,
-  RegtestClient,
-  script,
-  bufferUtils,
-};
+export { address, crypto, networks, payments, script, bufferUtils };
 
 export { Block } from './block.js';
 export { TaggedHashPrefix } from './crypto.js';
